requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
alluka==0.3.1
attrs==23.2.0
colorama==0.4.6
colorlog==6.8.2
frozenlist==1.4.1
hikari==2.0.0.dev126
hikari-lightbulb==2.3.5
hikari-miru==4.1.1
idna==3.7
multidict==6.0.5
python-dotenv==1.0.1
PyYAML==6.0.1
tinydb==4.8.0
typing_extensions==4.12.2
yarl==1.9.4